move StochasticDepthHparams
