+++ composer/algorithms/__init__.py
@@ -64,6 +64,7 @@ from composer.algorithms.hparams import LabelSmoothingHparams as LabelSmoothingH
 from composer.algorithms.hparams import MixUpHparams as MixUpHparams
 from composer.algorithms.hparams import RandAugmentHparams as RandAugmentHparams
 from composer.algorithms.hparams import SqueezeExciteHparams as SqueezeExciteHparams
+from composer.algorithms.hparams import StochasticDepthHparams as StochasticDepthHparams
 from composer.algorithms.label_smoothing import LabelSmoothing as LabelSmoothing
 from composer.algorithms.layer_freezing import LayerFreezing as LayerFreezing
 from composer.algorithms.layer_freezing import LayerFreezingHparams as LayerFreezingHparams
@@ -85,7 +86,6 @@ from composer.algorithms.squeeze_excite import SqueezeExcite as SqueezeExcite
 from composer.algorithms.squeeze_excite import SqueezeExcite2d as SqueezeExcite2d
 from composer.algorithms.squeeze_excite import SqueezeExciteConv2d as SqueezeExciteConv2d
 from composer.algorithms.stochastic_depth import StochasticDepth as StochasticDepth
-from composer.algorithms.stochastic_depth import StochasticDepthHparams as StochasticDepthHparams
 from composer.algorithms.swa import SWA as SWA
 from composer.algorithms.swa.hparams import SWAHparams as SWAHparams
