+++ composer/algorithms/stochastic_depth/__init__.py
@@ -3,7 +3,6 @@
 from composer.algorithms.stochastic_depth.sample_stochastic_layers import \
     SampleStochasticBottleneck as SampleStochasticBottleneck
 from composer.algorithms.stochastic_depth.stochastic_depth import StochasticDepth as StochasticDepth
-from composer.algorithms.stochastic_depth.stochastic_depth import StochasticDepthHparams as StochasticDepthHparams
 from composer.algorithms.stochastic_depth.stochastic_depth import apply_stochastic_depth as apply_stochastic_depth
 from composer.algorithms.stochastic_depth.stochastic_layers import StochasticBottleneck as StochasticBottleneck
